Treat errors as app insights exceptions
New constructor option `treatErrorsAsExceptions`.
When set to `true`, the transport will treat errors as app
insights exceptions for log levels >= `error` (defaults to false).

This allows you to see it clearly in the Azure Application Insights
instead of having to access trace information manually
and set up alerts based on the related metrics.

How it works:

* `winstonLogger.log('error', 'error message');`
   will trigger an app insights `trackException` with
   `Error('error message')` as argument

* `winstonLogger.log('error', new Error('error message'));`
   will trigger an app insights `trackException` with the
   Error object as argument

* `winstonLogger.log('error', 'error message', new Error('error message'));`
   will trigger an app insights `trackException`
   with the Error object as argument