Update log method + tests for new track* methods
Instead of relying on the defaultClient under the hood (behaviour induced by setting a key instead of specifying a client), we explicitly create a client for testing mocks. This avoids shared state behaviour where mocks may be called twice - speculatively this appears to be due to poor teardown of the applicationinsights client.